--- frontend/src/components/chat/config.js
@@ -1,10 +1,0 @@
-import pkg from '../../../package.json';
-
-export default {
-  development: {
-    endpoint: pkg.proxy
-  },
-  production: {
-    endpoint: window.location.hostname
-  }
-}
