+++ frontend/src/util/session_api_util.js
@@ -9,11 +9,9 @@ export const setAuthToken = token => {
 };
 
 export const signup = (userData) => {
-//////debugger
   return axios.post('/api/users/register', userData);
 };
 
 export const login = (userData) => {
-//////debugger
   return axios.post('/api/users/login', userData);
 };
